actually test vanillajs tests
Before, the file for vanillajs tests were not closed, making it test an
empty file which always passes.

Closing the file also reveals that the driverfile for vanillajs doesn't
properly work. The EXPORTS variable is initialized with old exports
before they're initialized in WebAssembly.instantiate. This makes it so
that the onInstantiation call gets exports that are mapped to the
__dummy function.

I'm not a 100% sure about the vanillajs driverfile fix. So any feedback
is welcome.