Run threaded tests on Node
This commit introduces the possibility for the test suite to run tests with Web Workers. This is done by introducing a polyfill for Workers as a CommonJS module.
The atomic tests have been modified to also run in ES6 module mode, so workers can be tested both on vanilla JS and with ES6 modules.